finetune/pyproject.toml
[project]
name = "hearsay-finetune"
version = "0.1.0"
description = "Hearsay transcription correction fine-tuning pipeline"
requires-python = ">=3.10"
dependencies = [
  "torch",
  "trl>=0.12.0",
  "peft>=0.7.0",
  "transformers>=4.45.0",
  "accelerate>=0.24.0",
  "huggingface_hub>=0.20.0",
  "datasets",
  "pyyaml",
  "sentencepiece",
  "pydantic>=2.0",
  "openai>=1.0.0",
  "dspy>=3.0",
  "mlx-lm>=0.20.0",
]

[dependency-groups]
dev = []
